+++ src/test/mixins/Configurable.sol
@@ -17,7 +17,7 @@ abstract contract Configurable {
         address payable assetRecipient,
         LSSVMPair.PoolType poolType,
         uint256 delta,
-        uint256 fee,
+        uint96 fee,
         uint256 spotPrice,
         uint256[] memory _idList,
         uint256 initialTokenBalance,
