+++ src/test/mixins/UsingETH.sol
@@ -35,7 +35,7 @@ abstract contract UsingETH is Configurable, RouterCaller {
         address payable assetRecipient,
         LSSVMPair.PoolType poolType,
         uint256 delta,
-        uint256 fee,
+        uint96 fee,
         uint256 spotPrice,
         uint256[] memory _idList,
         uint256,
